Fixes for CUDA ver. > 4 (sparse solver routines)
As far as I can tell from the online documentation the 'block' argument
has been removed in the Kernel.prepare call. I have changed the sparse
solver routines to reflect this.

In addition my system required that lists defining the grid and block in
in kernel execution calls were explicitly cast to tuples, if these were
passed in as list arguments boost_python would not able to make the
call.

To use the parse solver the file 'pycuda/sparse/pkt_build_cython.pyx'
must manually copied to the correct path within the python module
installed path, as the installer currently does not do this.